Refactor: Move Entities to domain package
