+++ src/main/java/toy/yogiyo/core/category/repository/CategoryRepository.java
@@ -1,7 +1,7 @@
 package toy.yogiyo.core.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import toy.yogiyo.core.category.Category;
+import toy.yogiyo.core.category.domain.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
