+++ src/main/java/toy/yogiyo/core/category/repository/CategoryShopRepository.java
@@ -1,7 +1,7 @@
 package toy.yogiyo.core.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import toy.yogiyo.core.category.CategoryShop;
+import toy.yogiyo.core.category.domain.CategoryShop;
 
 public interface CategoryShopRepository extends JpaRepository<CategoryShop, Long> {
