+++ src/main/java/toy/yogiyo/core/shop/dto/ShopDetailsResponse.java
@@ -2,7 +2,7 @@ package toy.yogiyo.core.shop.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import toy.yogiyo.core.shop.Shop;
+import toy.yogiyo.core.shop.domain.Shop;
 
 import java.util.List;
 import java.util.stream.Collectors;
