+++ src/main/java/toy/yogiyo/core/shop/repository/ShopRepository.java
@@ -1,7 +1,7 @@
 package toy.yogiyo.core.shop.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import toy.yogiyo.core.shop.Shop;
+import toy.yogiyo.core.shop.domain.Shop;
 
 public interface ShopRepository extends JpaRepository<Shop, Long> {
