+++ src/test/java/toy/yogiyo/core/shop/service/ShopServiceTest.java
@@ -12,8 +12,8 @@ import toy.yogiyo.common.exception.EntityNotFoundException;
 import toy.yogiyo.common.exception.FileIOException;
 import toy.yogiyo.common.file.ImageFileHandler;
 import toy.yogiyo.common.file.ImageFileUtil;
-import toy.yogiyo.core.shop.DeliveryPrice;
-import toy.yogiyo.core.shop.Shop;
+import toy.yogiyo.core.shop.domain.DeliveryPrice;
+import toy.yogiyo.core.shop.domain.Shop;
 import toy.yogiyo.core.shop.dto.DeliveryPriceDto;
 import toy.yogiyo.core.shop.dto.ShopRegisterRequest;
 import toy.yogiyo.core.shop.dto.ShopDetailsResponse;
